feat: 결과페이지 비교된 카드 mock server 적용
